Rewrite to super-simple static generator
This can then be run via a cron job to produce raw HTML that will be
served by NGINX, which is miles better at memory usage than Python.
Sanic was a fun experiment but it can't even serve a 100MB file without
using up all of the 512MB available memory.